HelloProject/backend/requirements.txt
django==4.2.7
django-ninja==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0
